+++ src/index.ts
@@ -704,6 +704,8 @@ export {
   spawnMissile,
 } from './b3d-launcher'
 export type { ProjectileOpts, MissileOpts, Impact } from './b3d-launcher'
+export { createMakers } from './make-mesh'
+export type { Makers, MakeOptions, MakeOwner } from './make-mesh'
 export { B3dTurret, b3dTurret } from './b3d-turret'
 export { B3dRadarBlip, b3dRadarBlip } from './b3d-radar-blip'
 export { B3dRadar, b3dRadar } from './b3d-radar'